pyproject.toml
[project]
name = "simple-django-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.38",
    "django>=5.2.6",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.1",
    "loguru>=0.7.3",
    "pillow>=11.3.0",
    "poethepoet>=0.37.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.12",
]
[tool.poe.tasks]
dev = "python manage.py runserver"

# Конфигурация Ruff
[tool.ruff]
# Длина строки
line-length = 88

# Python версия
target-version = "py312"

# Папки для проверки
include = ["*.py", "*.pyi"]

# Исключения
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "migrations",
    "*.pyc",
]

[tool.ruff.lint]
# Включаем правила
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings  
    "F",    # pyflakes
    "I",    # isort (импорты)
    "DJ",   # Django-specific rules
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "N",    # pep8-naming
]

# Отключаем некоторые правила
ignore = [
    "E501",   # line too long (пусть форматтер решает)
    "DJ01",   # Django model без __str__ (не всегда нужно)
]

# Настройки для Django
[tool.ruff.lint.per-file-ignores]
"manage.py" = ["E402"]  # импорты не в начале файла
"*/settings.py" = ["F405", "F403"]  # star imports в settings
"*/migrations/*.py" = ["E501", "F401"]  # длинные строки и неиспользуемые импорты в миграциях

[tool.ruff.lint.isort]
# Настройки сортировки импортов для Django
known-first-party = ["api", "simple_django_project"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.format]
# Настройки форматирования
quote-style = "double"
indent-style = "space"
line-ending = "auto"
